+++ lib/peg.d.ts
@@ -1160,16 +1160,6 @@ export interface BuildOptionsBase {
    */
   grammarSource?: any;
 
-  /**
-   * Selects between optimizing the generated parser for parsing speed (`"speed"`)
-   * or code size (`"size"`) (default: `"speed"`)
-   *
-   * @deprecated This feature was deleted in 1.2.0 release and has no effect anymore.
-   *             It will be deleted in 2.0.
-   *             Parser is always generated in the former `"speed"` mode
-   */
-  optimize?: "size" | "speed";
-
   /** Plugins to use */
   plugins?: Plugin[];